Fast API/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
numpy==1.24.0
pillow==9.4.0
tensorflow==2.12.0
opencv-python==4.7.0.72
